Create HuaweiMessagingDelegate class
