+++ stream-chat-android-pushprovider-huawei/src/main/java/io/getstream/chat/android/pushprovider/huawei/ChatHuaweiMessagingService.kt
@@ -3,12 +3,7 @@ package io.getstream.chat.android.pushprovider.huawei
 import android.util.Log
 import com.huawei.hms.push.HmsMessageService
 import com.huawei.hms.push.RemoteMessage
-import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.logger.ChatLogger
-import io.getstream.chat.android.client.models.Device
-import io.getstream.chat.android.client.models.PushMessage
-import io.getstream.chat.android.client.models.PushProvider
-import kotlin.jvm.Throws
 
 internal class ChatHuaweiMessagingService : HmsMessageService() {
     private val logger = ChatLogger.get("ChatHuaweiMessagingService")
@@ -16,30 +11,19 @@ internal class ChatHuaweiMessagingService : HmsMessageService() {
     override fun onMessageReceived(remoteMessage: RemoteMessage) {
         logger.logD("onMessageReceived(): $remoteMessage")
         try {
-            remoteMessage.takeIf { it.isValid() }
-                ?.toPushMessage()
-                ?.run { ChatClient.handlePushMessage(this) }
+            HuaweiMessagingDelegate.handleRemoteMessage(remoteMessage)
         } catch (exception: IllegalStateException) {
-            Log.e(TAG, "Error while handling remote message: ${exception.message}")
+            Log.e(TAG, "Error while handling remote message", exception)
         } finally {
             stopSelf()
         }
     }
 
-    override fun onNewToken(token: String?) {
+    override fun onNewToken(token: String) {
         try {
-            token?.takeUnless { it.isNullOrBlank() }
-                ?.run {
-                    ChatClient.setDevice(
-                        Device(
-                            token = token,
-                            pushProvider = PushProvider.HUAWEI,
-                        )
-                    )
-                }
-                ?: logger.logI("Empty token received from Huawei Push Kit")
+            HuaweiMessagingDelegate.registerHuaweiToken(token)
         } catch (exception: IllegalStateException) {
-            Log.e(TAG, "Error while handling remote message: ${exception.message}")
+            Log.e(TAG, "Error while registering Huawei Token", exception)
         }
     }
 
@@ -47,18 +31,3 @@ internal class ChatHuaweiMessagingService : HmsMessageService() {
         private const val TAG = "Chat:"
     }
 }
-
-@Throws(IllegalStateException::class)
-private fun RemoteMessage.toPushMessage() = when (isValid()) {
-    true -> PushMessage(
-        channelId = dataOfMap["channel_id"]!!,
-        messageId = dataOfMap["message_id"]!!,
-        channelType = dataOfMap["channel_type"]!!,
-    )
-    else -> throw IllegalStateException("RemoteMessage doesn't contains needed data")
-}
-
-private fun RemoteMessage.isValid() =
-    !dataOfMap["channel_id"].isNullOrBlank() &&
-        !dataOfMap["message_id"].isNullOrBlank() &&
-        !dataOfMap["channel_type"].isNullOrBlank()
